Move `VariableAxis` and `PinnedAxis` to `subgraph` module
