+++ apps/hash-graph/lib/graph/src/shared.rs
@@ -1,4 +1,3 @@
 pub mod identifier;
 pub mod interval;
 pub mod provenance;
-pub mod subgraph;
